+++ CHANGELOG.md
@@ -8,3 +8,4 @@ Initial release of vs code extension for easily creating openflow agent code or 
 - Initial release
 - Fix error when using non secure apiurl after choosing grpc
 - Fix project settings not getting populated to openflow
+- Fix issue with getting updates to package.json